support: Consolidate SRP for HAP implementations
SRP is 99% the same for MRP and Companion, so this change makes the code
generic enough to be shared by both protocols.

Relates to #655
